Previous commit to do 32 bit ratvec's needed a modification of this file as well, which I forgot.

+++ sources/utilities/arithmetic_fwd.h
@@ -18,14 +18,16 @@ namespace atlas {
 
 namespace arithmetic {
 
-  using Numer_t = long long int;
-  using Denom_t = unsigned long long int;
+  //  using Numer_t = long long int;
+  //  using Denom_t = unsigned long long int;
+  using Numer_t = int;
+  using Denom_t = unsigned int;
   template<typename I> class Rational;
   using RatNum = Rational<Numer_t>;
   using RatNumList = std::vector<RatNum>;
   class Split_integer;
-  class big_int; // defined in bigint.h but fiorward declared here
-  class big_rat; // defined in bigint.h but fiorward declared here
+  class big_int; // defined in bigint.h but forward declared here
+  class big_rat; // defined in bigint.h but forward declared here
 
 } // |namespace arithmetic|
